test(testCacheInterface): add
